Move TagsLogic into DockerAPI
